chore: shift `create_multi_index` to `lib.helpers`
