+++ unicon_backend/lib/common.py
@@ -1,23 +1,8 @@
 import re
-from collections import defaultdict
-from collections.abc import Callable
 from typing import Any
 
 from pydantic import BaseModel, model_validator
 from sqlmodel import MetaData, SQLModel
-
-
-def create_multi_index[T, K, V](
-    items: list[T],
-    key_fn: Callable[[T], K],
-    value_fn: Callable[[T], V],
-    filter_fn: Callable[[T], bool] = lambda _: True,
-) -> defaultdict[K, list[V]]:
-    """Create a one-to-many mapping index from a single list of items"""
-    index = defaultdict(list)
-    for item in filter(filter_fn, items):
-        index[key_fn(item)].append(value_fn(item))
-    return index
 
 
 def _camel_to_snake(name: str) -> str:
